+++ Sources/MacNotchIsland/Views/Panel/PanelView.swift
@@ -75,6 +75,7 @@ struct PanelView: View {
             switch HomeSection(rawValue: tab) ?? .home {
             case .home: HomeGridView()
             case .music: MusicSectionView(geometry: geometry)
+            case .controls: ControlsSectionView()
             case .today: TodaySectionView()
             case .windows: WindowsSectionView()
             case .shelf: ShelfSectionView(isDropTarget: isDropTarget)